requirements.txt
pillow==10.4.0
rich>=13.7.1
nbtlib==2.0.4
